Refactor prompt generation logic: move generate_llm_prompt to prompt_utils.py and update imports in coord2study and brain_insights modules.
