+++ coord2region/__init__.py
@@ -20,7 +20,10 @@ from .coord2study import (
     prepare_datasets,
     search_studies,
     get_studies_for_coordinate,
+)
+from .prompt_utils import (
     generate_llm_prompt,
+    generate_region_image_prompt,
 )
 
 # AI model interface
